Introduced osqp_api_constants.h
